feat: change the behavior of forgot-password
Send forgot password email to email address. The email will be sent in two case: First, the account's email is activated. Second, there is no account with this email has been activated, if so there will be two link corresponding to two roles to reset password. If you provides this function a email, in the database has two accounts for this email, and one account has been activated, then only account with role of activated account will be able to reset the password